move test definition initialization into separate file
